requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
networkx>=2.6.0
scikit-learn>=0.24.0
hdbscan>=0.8.0
powerlaw>=1.5.0
google-cloud-bigquery>=2.0.0
plotly>=5.0.0 